Use a simple server to create content and transfer files
Afte trying run the server in a thread and having problems
due to webkit bloking the ui, I implemented the server in another
package and start it in another process with subproces.
This solution solves the problem, and provide separation,
but can complicate other parts in the code, like collaboration.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>