Display a list of starred items in the journal
The server creates a json file with data about starred items
in the journal. The webkit display a html file with a javascript
to put the data in a table. Css is not working.
A download link allows download the file, but the client code
to download it is not included yet.
Is possible see the page to debug and download the file pointing
a browser to http://localhost:2500/web/index.html

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>