Upload metadata and preview
The filepicker is modified to create two files more, one with the preview,
and another with the metadata as a json dump.
The activity upload the 3 files, and the server put all the info in the
new object created in the journal.
One downside in this implementation is webkit show a message "3 files
selected", while the user selected one.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>