Prepare files to be served
As the web ask is tserver can't read the datastore because is in another thread
the activity prepare the files to be served in thier /instance/
directory before the server starts. Later we need implement
updating the files when the selection is changed.
A pending task is test uploading a file from another client.
The JournalManager class was moved from server.py to activity.py

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>