Upload & download the journal objects packaged
To upload or download a journal object, is zipped with a file with
the preview and another with the metadata in json format.
The save methods are used to package/unpackage when download or upload.
This solve a few issues:
* metadata & preview transference.
* simplify the code because only one file is transfered.
* the downloadmanager is always used, because webkit don't know
how manage the mime type used (solve error with images captured by
webkit)

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>