Use tornado as web server
In preparation to use websockets to notify the clients of changes,
changed the server to use tornado instead of BasicHttpServer.
The code is simplified and works fast.
I added the needed tornado files to the activity until we decide
if will be used as a global solution for web activities.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>